cleanup explorer link
