+++ app/src/utils/transfer.ts
@@ -1,4 +1,4 @@
-import { AssetHub, type Chain, KusamaAssetHub, type Token, type TokenAmount } from '@velocitylabs-org/turtle-registry'
+import type { Chain, Token, TokenAmount } from '@velocitylabs-org/turtle-registry'
 import { JsonRpcSigner } from 'ethers'
 import { toHex } from 'viem/utils'
 import type { Sender } from '@/hooks/useTransfer'
@@ -80,126 +80,6 @@ export const formatAmount = (amount: number, length: FormatLength = 'Short'): st
       maximumFractionDigits: length === 'Short' ? 1 : 3,
     }).format(amount)
   }
-}
-
-const removeURLSlash = (url: string) => {
-  if (url.length === 0) return url
-  const lastChar = url.charAt(url.length - 1)
-  if (lastChar === '/') {
-    return url.slice(0, -1)
-  } else {
-    return url
-  }
-}
-
-const EXPLORERS = {
-  // Ethereum
-  etherscan: 'https://etherscan.io/',
-
-  // Polkadot
-  subscan_polkadot: 'https://polkadot.subscan.io/',
-  subscan_polkadot_ah: 'https://assethub-polkadot.subscan.io/',
-  subscan_polkadot_bh: 'https://bridgehub-polkadot.subscan.io/',
-
-  // Kusama
-  subscan_kusama: 'https://kusama.subscan.io/',
-  subscan_kusama_ah: 'https://assethub-kusama.subscan.io/',
-  subscan_kusama_bh: 'https://bridgehub-kusama.subscan.io/',
-
-  // Chainflip
-  chainflip: 'https://scan.chainflip.io/',
-}
-
-export const getChainflipExplorerLink = (swap: StoredTransfer, chainflipSwapId?: string): string | undefined => {
-  if (chainflipSwapId) return `${removeURLSlash(EXPLORERS.chainflip)}/swaps/${chainflipSwapId}`
-  if (swap.uniqueTrackingId) return `${removeURLSlash(EXPLORERS.chainflip)}/channels/${swap.uniqueTrackingId}`
-}
-
-export function getExplorerLink(transfer: StoredTransfer): string | undefined {
-  const {
-    sourceChain: { network, chainId, walletType, name },
-    destChain,
-    sendResult: txHash,
-    sourceToken,
-    destinationToken,
-    sender,
-    id,
-    uniqueTrackingId,
-  } = transfer
-
-  // Chainflip explorer link - Not relying on network
-  const isChainflip = isChainflipSwap(transfer.sourceChain, destChain, sourceToken, destinationToken)
-  if (isChainflip) return getChainflipExplorerLink(transfer)
-
-  switch (network) {
-    case 'Ethereum': {
-      if (txHash) return `${removeURLSlash(EXPLORERS.etherscan)}/tx/${txHash}`
-
-      // Default Ethereum network explorer link:
-      return `${removeURLSlash(EXPLORERS.etherscan)}/address/${sender}`
-    }
-    case 'Polkadot': {
-      if (txHash) return `${removeURLSlash(EXPLORERS.subscan_polkadot_ah)}/extrinsic/${txHash}`
-
-      if (uniqueTrackingId) {
-        const path = getSubdomainPath(EXPLORERS.subscan_polkadot)
-        return `${removeURLSlash(EXPLORERS.subscan_polkadot)}/xcm_message/${path}-${uniqueTrackingId}`
-      }
-
-      if (walletType === 'SubstrateEVM') {
-        return getCustomExplorerLink(name, sender)
-      }
-
-      if (chainId === AssetHub.chainId) return `${removeURLSlash(EXPLORERS.subscan_polkadot_ah)}/extrinsic/${id}`
-
-      // Default Polkadot network explorer link:
-      return `${removeURLSlash(EXPLORERS.subscan_polkadot)}/account/${sender}?tab=xcm_transfer`
-    }
-    case 'Kusama': {
-      if (txHash) return `${removeURLSlash(EXPLORERS.subscan_kusama_ah)}/extrinsic/${txHash}`
-
-      if (uniqueTrackingId) {
-        const path = getSubdomainPath(EXPLORERS.subscan_kusama)
-        return `${removeURLSlash(EXPLORERS.subscan_kusama)}/xcm_message/${path}-${uniqueTrackingId}`
-      }
-
-      if (walletType === 'SubstrateEVM') {
-        return getCustomExplorerLink(name, sender)
-      }
-
-      if (chainId === KusamaAssetHub.chainId) return `${removeURLSlash(EXPLORERS.subscan_kusama_ah)}/extrinsic/${id}`
-
-      // Default Polkadot network explorer link:
-      return `${removeURLSlash(EXPLORERS.subscan_kusama)}/account/${sender}?tab=xcm_transfer`
-    }
-    default:
-      console.log(`Unsupported network: ${network}`)
-  }
-}
-
-/**
- * Extracts and returns the subdomain from a given URL.
- * For example,'https://polkadot.subscan.io/' input, returns 'polkadot'.
- * @param url - The URL from which the subdomain needs to be extracted. For example, "https://sub.example.com".
- * @returns The subdomain string from the URL.
- */
-const getSubdomainPath = (url: string) => {
-  // Generate a constructor URL. Example: 'https://polkadot.subscan.io/'
-  const parsedUrl = new URL(url)
-  // Filter hostname from URL: 'polkadot.subscan.io/'
-  const hostname = parsedUrl.hostname
-  // Split hostname & extract subdomain path: 'polkadot'
-  return hostname.split('.')[0]
-}
-
-/**
- * Generates the explorer link for SubstrateEVM walletType based chains. ex: Moonbmean, Mythos
- * @param name - The chain name.
- * @param sender - The sender address.
- * @returns The Subscan explorer link
- */
-const getCustomExplorerLink = (name: string, sender: string) => {
-  return `https://${name.toLowerCase()}.subscan.io/account/${sender}?tab=xcm_transfer`
 }
 
 export const txWasCancelled = (sender: Sender, error: unknown): boolean => {
